Smart Oven example added.
