+++ example/lib/pages/pages.dart
@@ -1,0 +1,2 @@
+export 'examples_page.dart';
+export 'smart_oven.dart';
